+++ client/app/src/main/java/com/example/witt/data/api/UserService.kt
@@ -6,7 +6,7 @@ import okhttp3.MultipartBody
 import okhttp3.RequestBody
 import retrofit2.http.*
 
-interface ProfileUploadService {
+interface UserService {
 
     @Multipart
     @PATCH("/api/user/me")
