+++ client/app/src/main/java/com/example/witt/data/source/remote/social_signin/SocialSignInDataSourceImpl.kt
@@ -1,6 +1,6 @@
 package com.example.witt.data.source.remote.social_signin
 
-import com.example.witt.data.api.SocialSignInService
+import com.example.witt.data.api.SignInService
 import com.example.witt.data.model.auth.request.SocialSignInRequest
 import com.example.witt.data.model.auth.response.SocialSignInResponse
 import com.example.witt.di.IoDispatcher
@@ -9,7 +9,7 @@ import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 class SocialSignInDataSourceImpl @Inject constructor(
-    private val socialSignInService: SocialSignInService,
+    private val socialSignInService: SignInService,
      @IoDispatcher private val coroutineDispatcher: CoroutineDispatcher
 ) : SocialSignInDataSource{
     override suspend fun signIn(socialSignInRequest: SocialSignInRequest): Result<SocialSignInResponse> =
