+++ client/app/src/main/java/com/example/witt/data/source/remote/user/TokenSignInDataSourceImpl.kt
@@ -1,6 +1,6 @@
 package com.example.witt.data.source.remote.token_signin
 
-import com.example.witt.data.api.TokenSignInService
+import com.example.witt.data.api.SignInService
 import com.example.witt.data.model.user.response.TokenResponse
 import com.example.witt.di.IoDispatcher
 import kotlinx.coroutines.CoroutineDispatcher
@@ -8,7 +8,7 @@ import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 class TokenSignInDataSourceImpl @Inject constructor(
-    private val userService: TokenSignInService,
+    private val userService: SignInService,
     @IoDispatcher private val coroutineDispatcher: CoroutineDispatcher
 ) : TokenSignInDataSource{
     override suspend fun tokenSignIn(): Result<TokenResponse> =
